# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"gettoken": {
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Login": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResPERSON": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"isAdmin": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"name",
					"email",
					"password",
					"isAdmin"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqPERSON": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"isAdmin": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"email",
					"password",
					"isAdmin"
				],
				"type": "object",
				"additionalProperties": false
			},
			"order": {
				"properties": {
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"type": {},
					"size": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"price",
					"type",
					"size"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResORDER": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"user": {
						"type": "string"
					},
					"orderItems": {
						"items": {
							"$ref": "#/components/schemas/order"
						},
						"type": "array"
					},
					"totalPrice": {
						"type": "number",
						"format": "double"
					},
					"isAccepted": {
						"type": "boolean"
					},
					"deliveredAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"user",
					"orderItems",
					"totalPrice",
					"isAccepted",
					"deliveredAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetORDER": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqORDER": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"user": {
						"type": "string"
					},
					"orderItems": {
						"items": {
							"$ref": "#/components/schemas/order"
						},
						"type": "array"
					},
					"totalPrice": {
						"type": "number",
						"format": "double"
					},
					"isAccepted": {
						"type": "boolean"
					},
					"deliveredAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"_id",
					"user",
					"orderItems",
					"totalPrice",
					"isAccepted",
					"deliveredAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_LeanDocument__LeanDocument_T__": {
				"properties": {},
				"type": "object"
			},
			"Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/_LeanDocument__LeanDocument_T__",
						"description": "This documents _id."
					},
					"__v": {
						"description": "This documents __v."
					},
					"id": {
						"description": "The string version of this documents _id."
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_": {
				"$ref": "#/components/schemas/Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"LeanDocument_this_": {
				"$ref": "#/components/schemas/Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_"
			},
			"IORDER": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"DeleteORDER": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResWAITER": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"isWaiter": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"name",
					"email",
					"password",
					"isWaiter"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqWAITER": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"isWaiter": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"email",
					"password",
					"isWaiter"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Menu": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"type",
					"price"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "Task",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/Person/loginPerson": {
			"post": {
				"operationId": "Getloginn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/gettoken"
								}
							}
						}
					}
				},
				"tags": [
					"person"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Login"
							}
						}
					}
				}
			}
		},
		"/Person/savePerson": {
			"post": {
				"operationId": "SavePerson",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResPERSON"
								}
							}
						}
					}
				},
				"tags": [
					"person"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqPERSON"
							}
						}
					}
				}
			}
		},
		"/Order/getorder": {
			"post": {
				"operationId": "Getadmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResORDER"
								}
							}
						}
					}
				},
				"tags": [
					"order"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetORDER"
							}
						}
					}
				}
			}
		},
		"/Order/saveOrder": {
			"post": {
				"operationId": "Saveadmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResORDER"
								}
							}
						}
					}
				},
				"tags": [
					"order"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/order"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/Order/updateorder": {
			"put": {
				"operationId": "UpdateAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResORDER"
								}
							}
						}
					}
				},
				"tags": [
					"order"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqORDER"
							}
						}
					}
				}
			}
		},
		"/Order/deleteorder": {
			"delete": {
				"operationId": "Deletadmin",
				"responses": {
					"200": {
						"description": "Admin deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IORDER"
								}
							}
						}
					}
				},
				"tags": [
					"order"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteORDER"
							}
						}
					}
				}
			}
		},
		"/Order/getorderlist": {
			"post": {
				"operationId": "GetadminList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResORDER"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"order"
				],
				"security": [],
				"parameters": []
			}
		},
		"/Waiter/saveWaiter": {
			"post": {
				"operationId": "Savewaiter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResWAITER"
								}
							}
						}
					}
				},
				"tags": [
					"waiter"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqWAITER"
							}
						}
					}
				}
			}
		},
		"/Waiter/getmenu": {
			"post": {
				"operationId": "Menu",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Menu"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"waiter"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/Waiter/loginWaiter": {
			"post": {
				"operationId": "Getloginn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/gettoken"
								}
							}
						}
					}
				},
				"tags": [
					"waiter"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Login"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}